+++ client/src/pages/Login.js
@@ -45,7 +45,7 @@ const Login = (props) => {
             {data ? (
               <p>
                 Success! You may now head{' '}
-                <Link to="/">back to the homepage.</Link>
+                <Link to="/interviews">back to the homepage.</Link>
               </p>
             ) : (
               <main className="flex-row justify-center mt-4">
